v4.0.0 with Major Changes (#2)
# What's the matter?

Up to version 3.x, the Postgres driver had a fatal problem. When a session-level lock was acquired in a transaction, it was not always possible to release the lock during the `finally` blocks; **once Postgres encountered an error in a transaction, it would kill all continuing operations with an error.**

Two approaches were adopted to solve this problem:

### Monitoring for commit/rollback events (including rollbacks to savepoints)

Even if the lock release process fails, it can be recovered by performing commit/rollback operations. When the event is detected, we will retry the lock release process.

### Issue savepoints on `withLocking()` calls

If a transaction has already been opened, a savepoint is created immediately after lock acquisition so that the lock can be released on the fly after recovering from an error that occurred within the callback function.

# Changes

- `AdvisoryLockServiceProvider` has been added; which is automatically discovered.
- `withLocking()` takes `ConnectionInterface` as a first argument.